Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -3,8 +3,8 @@
 
 
 int main() {
-    float a = 3.14;
+    int a = 3.14;
     
-    printf("You entered : %f", a );
+    printf("You entered : %d", a);
     return 0;
 }
